ClassifierTensorflow: Improve code in general.
Former-commit-id: fa77560a0709e73733276ff5003f8cadf2badadf

+++ ClassifierTensorFlow/src/classifier/prediction/losses/weightedBinaryCrossEntropy.py
@@ -6,7 +6,7 @@ from tensorflow.python.keras.losses import BinaryCrossentropy
 
 class WeightedBinaryCrossEntropy(BinaryCrossentropy):
 
-    def __init__(self, weights: List[int], from_logits: bool):
+    def __init__(self, weights: List[float], from_logits: bool):
         super().__init__(from_logits=from_logits)
         self.weights = weights
